refactor(core): Consolidate the update_workflow extraction back into one file (no-changelog)
The multi-file split (partial-update-schemas.ts, partial-update-steps.ts,
workflow-settings-guards.ts) is deferred to a deeper cross-tool refactor
the MCP team will own; for this ticket, every extracted function moves
back into update-workflow.tool.ts as a private (non-exported) top-level
declaration, in the same relative order as before: schemas & contract,
then settings guards, then pipeline steps, then the factory and handler.

Nothing else changes. createUpdateWorkflowTool is still the file's only
export, its 15-positional-argument signature is untouched, and every
function body, JSDoc comment, and already-documented contract carries
over unchanged — this is a pure move, not a rewrite. Verified no other
file imported from the three merged modules before deleting them.

No behaviour change: all 121 existing tests pass untouched, the handler
stays at cyclomatic complexity 9, and typecheck/lint are clean.
